7 UI improvements

+++ IncomeGoalTracker.Core/Services/Implementations/TrainingClassService.cs
@@ -79,6 +79,14 @@ namespace IncomeGoalTracker.Core.Services.Implementations
             try
             {
                 _logger.LogInformation($"Deleting Training Class {id}");
+                // Delete the Class CEU's first to avoid foreign key constraint violations
+                bool classCeuComplete = await _classCeuRepo.DeleteTrainingClassCeusAsync(id);
+                if (!classCeuComplete)
+                {
+                    _logger.LogError($"Failed to delete Class CEU's for Training Class {id}");
+                    return false;
+                }
+
                 bool complete = await _trainingClassRepo.DeleteTrainingClassAsync(id);
                 if(complete)
                 {
@@ -170,7 +178,7 @@ namespace IncomeGoalTracker.Core.Services.Implementations
                 }
                 else
                 {
-                    _logger.LogInformation("No training classes found");
+                    _logger.LogInformation($"No training classes found for Certificate Id {certificateId}");
                     return Enumerable.Empty<TrainingClass>();
                 }
             }
@@ -191,14 +199,51 @@ namespace IncomeGoalTracker.Core.Services.Implementations
 
         }
 
-        public async Task<bool> UpdateTrainingClassAsync(TrainingClass trainingClass)
+        public async Task<bool> UpdateTrainingClassAsync(TrainingClassView trainingClass)
         {
             try
             {
                 _logger.LogInformation($"Updating Training Class {trainingClass.Id}");
-                bool complete = await _trainingClassRepo.UpdateTrainingClassAsync(trainingClass);
+                var trainingClassModel = TrainingClassMapper.MapToModel(trainingClass);
+                bool complete = await _trainingClassRepo.UpdateTrainingClassAsync(trainingClassModel);
+                bool ceuComplete = false;
+
                 if (complete)
                 {
+                    foreach(var classCeu in trainingClass.ClassCeus)
+                    {
+                        ClassCeu classCeuModel = ClassCeuMapper.MapToModel(classCeu);
+                        classCeuModel.TrainingClassiD = trainingClass.Id;
+                        if(classCeuModel.Id == 0)
+                        {
+                            classCeuModel.TrainingClassiD = trainingClass.Id;
+                            int ceuId = await _classCeuRepo.AddClassCeuAsync(classCeuModel);
+                            if (ceuId > 0)
+                            {
+                                _logger.LogInformation($"Class CEU {classCeu.Id} added");
+                                ceuComplete = true;
+                            }
+                            else
+                            {
+                                _logger.LogError($"Failed to add Class CEU {classCeu.Id}");
+                                return false;
+                            }
+                        }
+                        else
+                        {
+                            ceuComplete = await _classCeuRepo.UpdateClassCeuAsync(classCeuModel);
+                        }
+                            
+                        if (ceuComplete)
+                        {
+                            _logger.LogInformation($"Class CEU {classCeu.Id} updated");
+                        }
+                        else
+                        {
+                            _logger.LogError($"Failed to update Class CEU {classCeu.Id}");
+                            return false;
+                        }
+                    }
                     _logger.LogInformation($"Training Class {trainingClass.Id} updated");
                     return true;
                 }
@@ -222,6 +267,48 @@ namespace IncomeGoalTracker.Core.Services.Implementations
                 _logger.LogError("************************");
                 throw;
             }
+        }
+
+        public async Task<TrainingClassView> GetTrainingClassByIdAsync(int id)
+        {
+            TrainingClass trainingClass = new TrainingClass();
+            TrainingClassView trainingClassView = new TrainingClassView();
+
+            try
+            {
+                trainingClass = await _trainingClassRepo.GetTrainingClassByIdAsync(id);
+                if (trainingClass is not null)
+                {
+                    IEnumerable<ClassCeu> ceus = await _classCeuRepo.GetClassCeuForTrainingClassAsync(trainingClass.Id);
+                    trainingClass.ClassCeus = ceus.ToList();
+
+
+                    trainingClassView = TrainingClassMapper.MapToView(trainingClass);
+                    foreach (var ceu in trainingClass.ClassCeus)
+                    {
+                        ClassCeuView classCeuView = ClassCeuMapper.MapToView(ceu);
+                        trainingClassView.ClassCeus.Add(classCeuView);
+                    }
+
+                }
+
+                return trainingClassView;
+            }
+            catch (SqlException ex)
+            {
+                _logger.LogError($"*****SQL Exception*****");
+                _logger.LogError(ex.Message);
+                _logger.LogError("************************");
+                throw;
+            }
+            catch (Exception ex)
+            {
+                _logger.LogError($"*****Exception*****");
+                _logger.LogError(ex.Message);
+                _logger.LogError("************************");
+                throw;
+            }
+
         }
     }
 }
